Move scene toggling resposibilites to SceneMachine.
